Tree Flow

Tree Flow is an R6 class that facilitates the creation and navigation of user-defined decision rules.

'Rules' are the prescriptive decision points observed looking visually at a typical flow diagram. The class provides methods for managing rules (i.e., adding, redefining, removing)

Each branch of a split is structured in an environment containing the following members:

$values

The values beloning to the current branch: ultimately a subset of the values of "root"

$name

The name associated with the current branch

$aliases

The alternative names associated with the current branch

$leaf.nodes

The distinct count of values:key

$rule

The quote expression to be applied to $values during tree population

$prev

A pointer to the previous branch ($prev)

$false

A pointer to the child branch containing values that evaluate as FALSE when the branch rule is applied

$true

A pointer to the child branch containing values that evaluate as TRUE when the branch rule is applied

$`#`

A pointer to the class environment: this allows certain global methods to be accessed via chaining with the '$' accessor.

Removing a branch (i.e., setting the branch to NULL) removes all child branches as well

Public fields

tree.name

The name of the tree

root

The root of the tree

rule.book

The list of decision rules

id

The index of the current branch

tree.key

The data.table key column name

graph

The igraph-package representation of the tree

Active bindings

total.branches

Count the number of branches

depth

Calculate tree depth

curnt

Returns the current branch

parent

Goto the parent branch: this is similar to $goto() except it only moves to the parent branch of the current branch

leaves

Return the content of the current branch object values

tree

Return the environment list for the tree

Methods

Public methods

tree.flow$new()

tree.flow$manage.rules()

tree.flow$add.branches()

tree.flow$set.branch()

tree.flow$del.branch()

tree.flow$goto()

tree.flow$list.branches()

tree.flow$make.decision()

tree.flow$populate()

tree.flow$plot()

tree.flow$clone()

Method new()

Initialize a new tree

Usage

tree.flow$new(init.values, init.key, name)

Arguments

init.values

The values with which the tree is to be populate: must be a list (each member of the list must be the same length) or named dimensional object

init.key

The element or column name from init.values to serve as the data.table key to be used for counting elements in a branch

name

The desired name of the decision support tree

Method manage.rules()

Register and manage decision rules by operating on the class member $rule.book, the list of rules for use by branches. It is up to the user to ensure that the rule expression is appropriate for the data type of the values used to initialize the class (argument init.values of class method $new())

Usage

tree.flow$manage.rules(..., rem = FALSE, rtype = "list", .debug = FALSE)

Arguments

...

dots_list: One or more lists with members name, def, and desc

rem

When TRUE, the rule is removed from the class member $rule.book (see section 'Details')

rtype

Denotes the type of structure for the argument this.rule; list, "nlist" (nested lists), and tabular (2D objects) are supported (see section 'Details').

.debug

(logical) When TRUE, additional execution information is produced

Details

this.rule

name: A label for the rule

def: A expression operating on branch values, or a function accepting branch value as its first argument. Either option must return a logical vector when evaluated or invoked.

desc: The plain-language description of the rule)

When more than one rule is supplied to '...', arguments rem and rtype, must be the same length as the number of supplied rules; otherwise, the first value in each argument is recycled to the length of '...'.

rtype

list : list(name = , def = , desc = )

nlist : list(list(name = , def = , desc = ), list(name = , def = , desc = ), ...); a maximum depth of two is supported

tabular : data{table, frame}(name = c(...), def = c(...), desc = c(...))

Returns

The class, invisibly

Method add.branches()

Add empty objects to the current branch's 'true' and 'false' objects. The class object pointing to the current branch is set to the branch being modified. Do not use this method to update branch values: this method will not overwrite an existing branch. Use methods $set.branch.name() and $set.branch.rule().

Usage

tree.flow$add.branches(
  branch = self$curnt,
  false.label = "false",
  true.label = "true",
  final.label = "END",
  rule = quote((TRUE)),
  aliases = NULL
)

Arguments

branch

The branch environment to which child branches are being added: either an object or the target branch's name.

false.label

The assignment object "name" in the "false" branch

true.label

The assignment object "name" in the "true" branch

final.label

The value to assign the terminating node indicator in the current branch

rule

A string indicating which registered rule is to be assigned to the branch before adding child

aliases

A character vector of additional names by which the branch may be understood (e.g., "name_2", "name_3", etc.)

Returns

The branch, invisibly

Method set.branch()

Replace the rule of the indicated branch

Usage

tree.flow$set.branch(
  branch = self$curnt,
  new.name = self$curnt$name,
  new.rule = self$curnt$rule
)

Arguments

branch

The target branch (by name or id) or, when NULL, the current branch

new.name

The desired name to use for the branch

new.rule

The desired rule to use for the branch

Returns

The branch, invisibly

Method del.branch()

Delete a branch: child branches are consequently removed

Usage

tree.flow$del.branch(branch = self$curnt, confirm = TRUE)

Arguments

branch

The target branch (by name or id) or, when NULL, the current branch

confirm

(logical | TRUE) When TRUE, deletion confirmation is provided

Returns

Invisibly, the class environment with the current branch set to the root

Method goto()

Jump to an arbitrary branch in the tree, setting the "current branch" pointer to the target branch

Usage

tree.flow$goto(name = NULL, id = private$envs[[name]], .verbose = FALSE)

Arguments

name

The name (not alias) of the target branch

id

The id value of the target branch

.verbose

When TRUE, the current branch's ID and name are echoed

Returns

The current branch, invisibly

Method list.branches()

Print the structure of the private list of branches

Usage

tree.flow$list.branches()

Returns

The list of branches, invisibly

Method make.decision()

Apply a branch's rule to it's stored values and populate downstream 'true' and 'false' child branches. The branch is returned invisibly.

Usage

tree.flow$make.decision(branch = self$curnt, .debug = FALSE)

Arguments

branch

A class branch object, defaulting to the currently pointed to branch

.debug

Print debugging information

Returns

The branch, invisibly

Method populate()

Iteratively populate the tree by rendering decisions only on branches that have non-NULL "true" and "false" values.

Usage

tree.flow$populate(branch = "root", .debug = FALSE)

Arguments

branch

The target branch (by name or id) or, when NULL, "root". Iteration begins at this point in the tree.

.debug

Print debugging information

Details

For a given branch, the values that map to TRUE are copied to the current branch's corresponding member $true and vice-versa. The values of the current branch are preserved for exploration branch-by-branch.

This is performed sequentially and recursively using a "first matching rule wins" method: this means that if the set of values for a branch could be mapped to TRUE in multiple branches of the same depth, it will only map to one branch due to the propagation of values described before.

Population begins at the root by default. Once populated a corresponding igraph-package member $graph is created

Returns

The class environment, invisibly

Method plot()

Creates an visualization of the tree using visIgraph. The branch list is converted into a graph via graph_from_literal

Usage

tree.flow$plot(
  ...,
  igraph.args = NULL,
  layout.args = list(visHierarchicalLayout = list()),
  refresh = FALSE
)

Arguments

...

dots_list: igraph expressions that will customize the graph before sending to visIgraph

igraph.args

A named list of arguments to send to visIgraph

layout.args

A named list of length 1L given as list("layout name" = <argument list>) (see https://datastorm-open.github.io/visNetwork/layout.html for details

refresh

(logical | FALSE) When TRUE, the preceding arguments are defaulted to existing values in $private$vizArgs

Details

For ..., the expressions are executed in the order given.

Use g as the symbol for the igraph object when needed (e.g., V(g) <- ... )

By default, vertices are supplied with a title and level attributes; edges with title and label attributes

Assignment to graph objects can use <-. The graph is automatically on each pass.

To updated the entire graph, use the following: g <<- ...

Library calls should be prefixed with "igraph::" unless the library has been loaded and attached

The class member $graph is not updated unless the self$graph <<- ... is used.

Method clone()

The objects of this class are cloneable with this method.

Usage

tree.flow$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.